fix large anonymous function in Formik --> onSubmit to make it more readable. Export validationSchema as a constant.
