+++ src/components/NewEventForm/NewEventForm.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Formik, Form } from 'formik';
 import { addDays, format, set, startOfDay } from 'date-fns';
-import * as Yup from 'yup';
 
+import { eventFormSchema } from './constants';
 import Button from '../inputs/Button';
 import FormikInputField from '../FormikFields/FormikInputField/FormikInputField';
 import FormikTextAreaField from '../FormikFields/FormikTextAreaField/FormikTextAreaField';
@@ -46,20 +46,19 @@ type EventFormProps = {
 } & EventDataInput;
 
 function NewEventForm({ onSubmit, ...rest }: EventFormProps) {
-  // Declartions
-  const now = new Date();
-  const tomorrow = addDays(startOfDay(now), 1);
-
   // Utilities
   const setHours = (date: Date, hours: number) =>
     set(date, { hours: hours, minutes: 0, seconds: 0, milliseconds: 0 });
   const parseDate = (date: string, time: string) => new Date(date + 'T' + time);
 
+  // Declartions
+  const now = new Date();
+  const tomorrow = addDays(startOfDay(now), 1);
   const eventDate = format(tomorrow, 'yyyy-MM-dd');
   const eventStartTime = format(setHours(tomorrow, 16), 'HH:mm');
   const eventEndTime = format(setHours(tomorrow, 17), 'HH:mm');
 
-  // Initiation of values
+  // Initiation of form values
   const initialValues = {
     name: rest.name || '',
     eventType: rest.eventType || '',
@@ -74,40 +73,28 @@ function NewEventForm({ onSubmit, ...rest }: EventFormProps) {
     eventStatus: rest.eventStatus || EVENT_STATUS_OPTIONS.Draft,
   };
 
-  // Validation
-  const eventFormSchema = Yup.object().shape({
-    name: Yup.string()
-      .max(100, 'Event name is too long')
-      .required('Event name is required'),
-    eventType: Yup.string()
-      .oneOf(EVENT_OPTIONS)
-      .required('Event type is required'),
-    date: Yup.date().required('Event date is required'),
-    startTime: Yup.string()
-      .required('Start time is required')
-      .test(
-        'Start Time',
-        'Start time cannot be later than end time',
-        function (startTime) {
-          const startTimeHour = parseInt(
-            this.parent.startTime.toString().substring(0, 2)
-          );
-          const endTimeHour = parseInt(
-            this.parent.endTime.toString().substring(0, 2)
-          );
-          return startTimeHour < endTimeHour;
-        }
-      ),
-    endTime: Yup.string().required('End time is required'),
-    description: Yup.string()
-      .max(5000, 'Description text is too long')
-      .required('Description of event is required'),
-    location: Yup.string().max(200, 'Location text is too long'),
-    imageUrl: Yup.string().url('URL not valid'),
-    eventStatus: Yup.string()
-      .oneOf([EVENT_STATUS_OPTIONS.Draft, EVENT_STATUS_OPTIONS.Published])
-      .required('Event status is required'),
-  });
+  // Helper function for Formik --> onSubmit
+  async function submitEventData(values: any, setSubmitting: any): Promise<void> {
+    var promise;
+    setSubmitting(true);
+    try {
+      promise = onSubmit({
+        id: rest.id,
+        name: values.name,
+        description: values.description,
+        startTime: parseDate(eventDate, values.startTime),
+        endTime: parseDate(eventDate, values.endTime),
+        imageUrl: values.imageUrl,
+        imageAlt: values.imageAlt,
+        location: values.location,
+        eventType: values.eventType,
+        eventStatus: values.eventStatus,
+        teamsUrl: values.teamsUrl,
+      });
+    } catch (error) {}
+    setSubmitting(false);
+    return promise;
+  }
 
   return (
     <div>
@@ -115,26 +102,7 @@ function NewEventForm({ onSubmit, ...rest }: EventFormProps) {
         initialValues={initialValues}
         validationSchema={eventFormSchema}
         onSubmit={async (values, { setSubmitting }) => {
-          await new Promise((r) => {
-            setSubmitting(true);
-
-            try {
-              onSubmit({
-                id: rest.id,
-                name: values.name,
-                description: values.description,
-                startTime: parseDate(eventDate, values.startTime),
-                endTime: parseDate(eventDate, values.endTime),
-                imageUrl: values.imageUrl,
-                imageAlt: values.imageAlt,
-                location: values.location,
-                eventType: values.eventType,
-                eventStatus: values.eventStatus,
-                teamsUrl: values.teamsUrl,
-              });
-            } catch (error) {}
-          });
-          setSubmitting(false);
+          await submitEventData(values, setSubmitting);
         }}
       >
         <EventFormWrapper>
